// src/app/models/enums/MapperOgFagsakerActionKeys.ts
export enum MapperOgFagsakerActionKeys {

    MAPPER_SET = 'PUNCH_MAPPER_SET',
    MAPPER_LOAD = 'PUNCH_MAPPER_LOAD',
    MAPPER_REQUEST_ERROR = 'PUNCH_MAPPER_REQUEST_ERROR',

    MAPPE_OPEN = 'PUNCH_MAPPE_OPEN',
    MAPPE_CLOSE = 'PUNCH_MAPPE_CLOSE',
    MAPPE_CHOOSE = 'PUNCH_MAPPE_CHOOSE',
    MAPPE_UNDO_CHOICE = 'PUNCH_MAPPE_UNDO_CHOICE',

    FAGSAKER_SET = 'PUNCH_FAGSAKER_SET',
    FAGSAKER_LOAD = 'PUNCH_FAGSAKER_LOAD',
    FAGSAKER_REQUEST_ERROR = 'PUNCH_FAGSAKER_REQUEST_ERROR',

    FAGSAK_OPEN = 'PUNCH_FAGSAK_OPEN',
    FAGSAK_CLOSE = 'PUNCH_FAGSAK_CLOSE',

    MAPPE_CREATE_REQUEST = 'PUNCH_MAPPE_CREATE_REQUEST',
    MAPPE_CREATE_ERROR = 'PUNCH_MAPPE_CREATE_ERROR',
    MAPPE_CREATE_SUCCESS = 'PUNCH_MAPPE_CREATE_SUCCESS',

    MAPPEID_RESET = 'PUNCH_MAPPEID_RESET'
}